Restructure Nori-30M as a version within the nori package (review feedback)
Per @LennartPurucker's review: treat Nori-30M as a version/subclass of the existing
Nori code rather than a separate top-level model. Folds the standalone models/nori30m/
package into models/nori/:

- models/nori/model.py: Nori30MModel(NoriModel) added alongside NoriModel.
- models/nori/{info,hpo,__init__}.py: nori30m_method_metadata / nori30m_info / gen_nori30m
  added alongside the base (registry auto-discovers the second ModelInfo).
- models/__init__.py + contexts/tabarena/methods.py: imports repointed to models.nori.
- pyproject.toml: dropped the separate nori30m extra; bumped the nori extra to
  synthefy-nori>=0.10.0 (the model= selector, needed for Nori-30M).
- Removed packages/tabarena/src/tabarena/models/nori30m/.

Still a distinct "Nori-30M" leaderboard entry (does not supersede the base Nori).
Registry / lazy-import / method-metadata tests pass (31).
