+++ packages/tabarena/src/tabarena/models/nori/__init__.py
@@ -1,6 +1,18 @@
 from __future__ import annotations
 
-from tabarena.models.nori.hpo import gen_nori
-from tabarena.models.nori.info import nori_info, nori_method_metadata
+from tabarena.models.nori.hpo import gen_nori, gen_nori30m
+from tabarena.models.nori.info import (
+    nori30m_info,
+    nori30m_method_metadata,
+    nori_info,
+    nori_method_metadata,
+)
 
-__all__ = ["gen_nori", "nori_info", "nori_method_metadata"]
+__all__ = [
+    "gen_nori",
+    "gen_nori30m",
+    "nori30m_info",
+    "nori30m_method_metadata",
+    "nori_info",
+    "nori_method_metadata",
+]
