+++ packages/tabarena/src/tabarena/models/nori/hpo.py
@@ -1,10 +1,16 @@
 from __future__ import annotations
 
-from tabarena.models.nori.model import NoriModel
+from tabarena.models.nori.model import Nori30MModel, NoriModel
 from tabarena.utils.config_utils import ConfigGenerator
 
 gen_nori = ConfigGenerator(
     model_cls=NoriModel,
+    search_space={},
+    manual_configs=[{}],
+)
+
+gen_nori30m = ConfigGenerator(
+    model_cls=Nori30MModel,
     search_space={},
     manual_configs=[{}],
 )
