--- packages/tabarena/src/tabarena/models/nori30m/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-from tabarena.models.nori30m.hpo import gen_nori30m
-from tabarena.models.nori30m.info import nori30m_info, nori30m_method_metadata
-
-__all__ = ["gen_nori30m", "nori30m_info", "nori30m_method_metadata"]
